Split UmapFeatureSettings into DimensionalityReductionSettings and UmapFeatureSettings
