+++ src/main/java/org/mastodon/mamut/feature/dimensionalityreduction/umap/UmapFeatureSettings.java
@@ -28,6 +28,8 @@
  */
 package org.mastodon.mamut.feature.dimensionalityreduction.umap;
 
+import org.mastodon.mamut.feature.dimensionalityreduction.DimensionalityReductionSettings;
+
 /**
  * Settings for the UMAP feature.
  * <br>
@@ -35,60 +37,44 @@ package org.mastodon.mamut.feature.dimensionalityreduction.umap;
  * <ul>
  *     <li>the number of neighbors to consider in the umap algorithm</li>
  *     <li>the minimum distance between points</li>
- *     <li>whether to standardize features</li>
- *     <li>the number of output dimensions</li>
  * </ul>
  */
-public class UmapFeatureSettings
+public class UmapFeatureSettings extends DimensionalityReductionSettings
 {
-	public static final int DEFAULT_NUMBER_OF_OUTPUT_DIMENSIONS = 2;
-
 	public static final int DEFAULT_NUMBER_OF_NEIGHBORS = 15;
 
 	public static final double DEFAULT_MINIMUM_DISTANCE = 0.1d;
-
-	public static final boolean DEFAULT_STANDARDIZE_FEATURES = true;
-
-	private int numberOfOutputDimensions;
 
 	private int numberOfNeighbors;
 
 	private double minimumDistance;
 
-	private boolean standardizeFeatures;
-
 	/**
 	 * Constructor with default values.
 	 * Default values are:
 	 * <ul>
-	 *     <li>number of dimensions: {@value DEFAULT_NUMBER_OF_OUTPUT_DIMENSIONS}</li>
 	 *     <li>number of neighbors: {@value DEFAULT_NUMBER_OF_NEIGHBORS}</li>
 	 *     <li>minimum distance: {@value DEFAULT_MINIMUM_DISTANCE}</li>
-	 *     <li>standardize features: {@value DEFAULT_STANDARDIZE_FEATURES}</li>
 	 * </ul>
 	 */
 	public UmapFeatureSettings()
 	{
-		this( DEFAULT_NUMBER_OF_OUTPUT_DIMENSIONS, DEFAULT_NUMBER_OF_NEIGHBORS, DEFAULT_MINIMUM_DISTANCE, DEFAULT_STANDARDIZE_FEATURES );
+		this( DEFAULT_NUMBER_OF_NEIGHBORS, DEFAULT_MINIMUM_DISTANCE );
 	}
 
-	/**
-	 * Constructor with number of neighbors.
-	 *
-	 * @param numberOfOutputDimensions the number of neighbors to consider for relative movement.
-	 */
+	public UmapFeatureSettings( final int numberOfNeighbors, final double minimumDistance )
+	{
+		super();
+		this.numberOfNeighbors = numberOfNeighbors;
+		this.minimumDistance = minimumDistance;
+	}
+
 	public UmapFeatureSettings( final int numberOfOutputDimensions, final int numberOfNeighbors, final double minimumDistance,
 			final boolean standardizeFeatures )
 	{
-		this.numberOfOutputDimensions = numberOfOutputDimensions;
+		super( numberOfOutputDimensions, standardizeFeatures );
 		this.numberOfNeighbors = numberOfNeighbors;
 		this.minimumDistance = minimumDistance;
-		this.standardizeFeatures = standardizeFeatures;
-	}
-
-	public int getNumberOfOutputDimensions()
-	{
-		return numberOfOutputDimensions;
 	}
 
 	public int getNumberOfNeighbors()
@@ -101,16 +87,6 @@ public class UmapFeatureSettings
 		return minimumDistance;
 	}
 
-	public boolean isStandardizeFeatures()
-	{
-		return standardizeFeatures;
-	}
-
-	public void setNumberOfOutputDimensions( final int numberOfOutputDimensions )
-	{
-		this.numberOfOutputDimensions = numberOfOutputDimensions;
-	}
-
 	public void setNumberOfNeighbors( final int numberOfNeighbors )
 	{
 		this.numberOfNeighbors = numberOfNeighbors;
@@ -121,15 +97,10 @@ public class UmapFeatureSettings
 		this.minimumDistance = minimumDistance;
 	}
 
-	public void setStandardizeFeatures( final boolean standardizeFeatures )
-	{
-		this.standardizeFeatures = standardizeFeatures;
-	}
-
 	@Override
 	public String toString()
 	{
-		return "UmapFeatureSettings{" + "numberOfOutputDimensions=" + numberOfOutputDimensions + ", numberOfNeighbors=" + numberOfNeighbors
-				+ ", minimumDistance=" + minimumDistance + ", standardizeFeatures=" + standardizeFeatures + '}';
+		return super.toString() + ", UmapFeatureSettings{numberOfNeighbors=" + numberOfNeighbors + ", minimumDistance=" + minimumDistance
+				+ '}';
 	}
 }
